When moving primary axes together with secondary axes, the secondary axes' acceleration config (i.e. delta_acceleration, epsilon_acceleration, zeta_acceleration) is not respected.
When the secondary axes' delta is large, this results in exceeding the acceleration the axis can take. This may result in step losses or damage to your tools or work piece.

The shortcoming is known to the authors, as is documented by the TODO in the current code:

// TODO we may need to do all of them, check E won't limit XYZ.. it does on long E moves, but not checking it could exceed the E acceleration.

See:
https://github.com/markmaker/Smoothieware/blob/17ff1036f00cd9cfdca041eda7c997339c2fc7e8/src/modules/robot/Robot.cpp#L1308

As the secondary axes (extruders etc.) do not affect the primary toolpath they are also not part of the Euclidean distance sqrt(sum of squares) of the move. However if the primary axes do not move at all (auxiliary move), another distance value must be calculated. Currently another Euclidean distance over the secondary axes is taken.

However for purely auxiliary moves, it seems better to use the maximum over the absolute deltas and not the Euclidean distance.

Example 1: for a 3D Printer with dual extrusion, printing two objects in parallel (i.e. BCN3D Sigma), feed rate and acceleration should NOT be scaled down by the Euclidean distance over both extrusions as this would unexpectedly scale down feed rate and acceleration by 1/sqrt(2).

Example 2: a pick and place machine with two nozzles has two rotation axes for placing parts at the correct angle. The two rotations are completely independent, so it does not make sense to calculate the Euclidean distance over the two angles and then scale down speed and acceleration of both. The same is even more evident, if one axis serves a completely different purpose, such as a part feeder, conveyor etc.

However...

Example 3: A 4-axis CNC with a rotation axis will need to compile with PAXIS=4 resulting in the macro N_PRIMARY_AXIS=4. This is because the rotation axis, through rotation of the tool tip or work piece, indeed does add linearly (for a given short moment of time dt) to the Euclidean distance along the toolpath, together with the XYZ motion (the G code will have to issue fine grained way points and feed rates to compensate for the wildly varying radii in effect here). In this case axis 4 is a primary axis and will by definition never be part of an auxiliary move. The maximum distance determination - correctly - does not apply.

By making this distinction correctly, it seems the change to maximum absolute distance is justified.