Fixed junction deviation for mixed primary/auxiliary axes.
1. Respect auxiliary axes

When primary and auxiliary axes were mixed the junction deviation was calculated without the auxiliary axes.
So when primary axes had a relatively straight junction the planner will speed through the junction.
Consequently the auxiliary axes were overwhelmed, stalling the steppers.

This is now fixed by taking the cosinus(theta) over all the axes. The unit vector now contains all the axes
and the distance to norm it is the overall multi-dimensional Euclidean distance.

Junction deviation is disabled if a primary move (including a mixed one) follows a purely auxiliary move and
vice versa. The motion has different units and can therefore not be compared. However as long as the segments
remain in the same domain (primary or auxiliary) they can be optimized with junction deviation.

The "z only" junction deviation is only chosen, if not auxiliary axes are in the move. Formerly it only looked
at X, Y, Z.

2. Symmetric junction deviation and acceleration limits

When the set of axes changes from before the junction to after the junction i.e. some axes stop at the junction,
the admissible acceleration may change. Therefore the acceleration of both before and after the junction must
be respected. The change does that, by taking the minimum acceleration of both. See more details in the commend
of the code.